DynamoDB/requirements.txt
streamlit
boto3
pandas
pillow
